+++ src/components/leaderboard/card-list/CardList.module.scss
