+++ src/components/shared/icon/Icon.module.scss
