+++ src/container/main-menu/MainMenu.module.scss
@@ -13,11 +13,11 @@
     color: #fff;
   }
   h1 {
-    font-size: clamp(1.9rem, 4.1vw, 3rem);
+    font-size: clamp(2rem, 5.1vw, 3.2rem);
     margin-top: 0px;
     padding: 0 10px;
   }
   p {
-    font-size: clamp(1rem, 1.5vw, 2rem);
+    font-size: clamp(1.2rem, 1.6vw, 1.8rem);
   }
 }
