+++ src/container/routes/PrivateRoutes.jsx
@@ -1,6 +1,6 @@
 import { Outlet, Navigate } from 'react-router-dom';
 import { useAuthContext } from 'context/auth-context/AuthContainer';
-import { LOGIN_ROUTE } from 'container/routes/ConstantsZX';
+import { LOGIN_ROUTE } from 'container/routes/Constants';
 const PrivateRoutes = () => {
   const { isAuthenticated } = useAuthContext();
   return isAuthenticated ? <Outlet /> : <Navigate to={LOGIN_ROUTE} />;
